@@ -16,87 +16,87 @@
   },
   "dependencies": {
     "@ant-design/icons": "^4.2.2",
+    "@apollo/client": "^3.3.21",
     "@apollo/react-hooks": "^4.0.0",
-    "@ethersproject/address": "^5.0.5",
-    "@ethersproject/bignumber": "^5.0.8",
-    "@ethersproject/bytes": "^5.0.5",
-    "@ethersproject/contracts": "^5.0.5",
-    "@ethersproject/providers": "^5.5.6",
-    "@ethersproject/solidity": "^5.0.9",
-    "@ethersproject/units": "^5.0.6",
+    "@portis/web3": "^4.0.5",
     "@ramp-network/ramp-instant-sdk": "^2.2.0",
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.8",
     "@uniswap/sdk": "^3.0.3",
     "@uniswap/v2-periphery": "^1.1.0-beta.0",
-    "@walletconnect/web3-provider": "^1.6.5",
-    "antd": "^4.7.0",
+    "@walletconnect/web3-provider": "^1.5.2",
+    "antd": "4.16.0",
     "apollo-boost": "^0.4.9",
     "apollo-client": "^2.6.10",
     "apollo-utilities": "^1.3.4",
-    "axios": "^0.20.0",
-    "bip39": "^3.0.4",
+    "arb-ts": "^0.0.18",
+    "authereum": "^0.1.14",
+    "axios": "^0.25.0",
+    "bl": "^5.0.0",
     "bnc-notify": "^1.5.0",
-    "burner-provider": "^1.0.38",
     "dotenv": "^8.2.0",
     "eth-hooks": "2.3.16",
-    "ethereumjs-wallet": "^1.0.2",
-    "ethers": "^5.0.31",
-    "graphiql": "^1.0.5",
+    "ethers": "^5.4.1",
+    "fortmatic": "^2.2.1",
+    "graphiql": "^1.4.7",
     "graphql": "^15.3.0",
+    "ipfs-http-client": "^55.0.0",
     "isomorphic-fetch": "^3.0.0",
     "node-watch": "^0.7.1",
     "postcss": "^8.2.6",
     "qrcode.react": "^1.0.0",
-    "react": "^16.14.0",
+    "react": "^17.0.2",
     "react-blockies": "^1.4.1",
     "react-css-theme-switcher": "^0.2.2",
-    "react-dom": "^16.14.0",
+    "react-dom": "^17.0.2",
     "react-qr-reader": "^2.2.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.3",
+    "react-scripts": "4.0.0",
+    "searchico": "^1.2.1",
+    "walletlink": "^2.1.5",
     "web3modal": "^1.9.1"
   },
   "devDependencies": {
+    "@apollo/client": "^3.3.21",
     "@testing-library/dom": "^6.12.2",
     "@types/react": "^16.9.19",
     "autoprefixer": "^10.2.4",
     "chalk": "^4.1.0",
-    "eslint": "^7.5.0",
-    "eslint-config-airbnb": "^18.2.0",
-    "eslint-config-prettier": "^6.11.0",
-    "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-prettier": "^3.1.4",
-    "eslint-plugin-react": "^7.22.0",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-prettier": "^4.0.0",
     "gulp": "^4.0.2",
     "gulp-csso": "^4.0.1",
     "gulp-debug": "^4.0.0",
     "gulp-less": "^4.0.1",
     "gulp-postcss": "^9.0.0",
-    "ipfs-http-client": "^45.0.0",
     "less-plugin-npm-import": "^2.1.0",
     "prettier": "^2.0.5",
     "s3-folder-upload": "^2.3.1",
     "surge": "^0.21.5"
   },
   "eslintConfig": {
-    "extends": "react-app"
+    "extends": [
+      "react-app",
+      "plugin:prettier/recommended"
+    ],
+    "rules": {
+      "prettier/prettier": "warn"
+    }
   },
   "scripts": {
     "build": "react-scripts build",
     "eject": "react-scripts eject",
+    "prestart": "node ./scripts/create_contracts.js",
     "start": "react-scripts start",
     "test": "react-scripts test",
     "lint": "eslint --config ./.eslintrc.js --ignore-path ./.eslintignore ./src/**/*",
     "ipfs": "node ./scripts/ipfs.js",
-    "surge": "surge ./build",
+    "surge": "cp build/index.html build/200.html && surge ./build",
     "s3": "node ./scripts/s3.js",
     "ship": "yarn surge",
     "theme": "npx gulp less",
-    "watch": "node ./scripts/watch.js"
+    "watch": "node ./scripts/watch.js",
+    "prettier": "npx prettier --write . '!(node_module|build)/**/*'"
   }
 }
