@@ -1,14 +1,16 @@
-/* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable jsx-a11y/no-static-element-interactions */
-/* eslint-disable jsx-a11y/accessible-emoji */
-import { BigNumber } from "@ethersproject/bignumber";
 import { Button, Col, Divider, Input, Row, Tooltip } from "antd";
 import React, { useState } from "react";
 import Blockies from "react-blockies";
-import { Transactor } from "../../helpers";
-import tryToDisplay from "./utils";
 
-const { utils } = require("ethers");
+import { Transactor } from "../../helpers";
+import { tryToDisplay, tryToDisplayAsText } from "./utils";
+
+const { utils, BigNumber } = require("ethers");
+
+const getFunctionInputKey = (functionInfo, input, inputIndex) => {
+  const name = input?.name ? input.name : "input_" + inputIndex + "_";
+  return functionInfo.name + "_" + name + "_" + input.type;
+};
 
 export default function FunctionForm({ contractFunction, functionInfo, provider, gasPrice, triggerRefresh }) {
   const [form, setForm] = useState({});
@@ -17,9 +19,8 @@ export default function FunctionForm({ contractFunction, functionInfo, provider,
 
   const tx = Transactor(provider, gasPrice);
 
-  let inputIndex = 0;
-  const inputs = functionInfo.inputs.map(input => {
-    const key = functionInfo.name + "_" + input.name + "_" + input.type + "_" + inputIndex++;
+  const inputs = functionInfo.inputs.map((input, inputIndex) => {
+    const key = getFunctionInputKey(functionInfo, input, inputIndex);
 
     let buttons = "";
     if (input.type === "bytes32") {
@@ -159,6 +160,12 @@ export default function FunctionForm({ contractFunction, functionInfo, provider,
     inputs.push(txValueInput);
   }
 
+  const handleForm = returned => {
+    if (returned) {
+      setForm({});
+    }
+  };
+
   const buttonIcon =
     functionInfo.type === "call" ? (
       <Button style={{ marginLeft: -32 }}>Read📡</Button>
@@ -178,9 +185,8 @@ export default function FunctionForm({ contractFunction, functionInfo, provider,
             style={{ width: 50, height: 30, margin: 0 }}
             type="default"
             onClick={async () => {
-              let innerIndex = 0;
-              const args = functionInfo.inputs.map(input => {
-                const key = functionInfo.name + "_" + input.name + "_" + input.type + "_" + innerIndex++;
+              const args = functionInfo.inputs.map((input, inputIndex) => {
+                const key = getFunctionInputKey(functionInfo, input, inputIndex);
                 let value = form[key];
                 if (input.baseType === "array") {
                   value = JSON.parse(value);
@@ -196,18 +202,27 @@ export default function FunctionForm({ contractFunction, functionInfo, provider,
 
               let result;
               if (functionInfo.stateMutability === "view" || functionInfo.stateMutability === "pure") {
-                const returned = await contractFunction(...args);
-                result = tryToDisplay(returned);
+                try {
+                  const returned = await contractFunction(...args);
+                  handleForm(returned);
+                  result = tryToDisplayAsText(returned);
+                } catch (err) {
+                  console.error(err);
+                }
               } else {
                 const overrides = {};
                 if (txValue) {
                   overrides.value = txValue; // ethers.utils.parseEther()
+                }
+                if (gasPrice) {
+                  overrides.gasPrice = gasPrice;
                 }
                 // Uncomment this if you want to skip the gas estimation for each transaction
                 // overrides.gasLimit = hexlify(1200000);
 
                 // console.log("Running with extras",extras)
                 const returned = await tx(contractFunction(...args, overrides));
+                handleForm(returned);
                 result = tryToDisplay(returned);
               }
 
