@@ -3,7 +3,14 @@ const fs = require("fs");
 
 const directoryName = "build";
 
-const BUCKETNAME = "punkwallet.io"; // <<---- SET YOUR BUCKET NAME AND CREATE aws.json ** see below vvvvvvvvvv
+const BUCKETNAME = "multisig.lol"; // <<---- SET YOUR BUCKET NAME AND CREATE aws.json ** see below vvvvvvvvvv
+
+
+const invalidation = {
+ awsDistributionId: "E3MT2OEGZAOJTZ",
+ awsInvalidationPath: "/*"
+}
+
 
 if (!BUCKETNAME) {
   console.log("☢️   Enter a bucket name in packages/react-app/scripts/s3.js ");
@@ -20,7 +27,6 @@ try {
   );
   process.exit(1);
 }
-console.log("credentials", credentials);
 
 credentials.bucket = BUCKETNAME;
 
@@ -30,10 +36,57 @@ const options = {
   useIAMRoleCredentials: false,
 };
 
-// optional cloudfront invalidation rule
-const invalidation = {
- awsDistributionId: "E3D4GB8Y5ZDQB0",
- awsInvalidationPath: "/*"
-}
+/////////////
+///////////// First, let's automatically create the bucket if it doesn't exist...
+/////////////
 
-s3FolderUpload(directoryName, credentials, options  , invalidation );
+var AWS = require('aws-sdk');
+// Load credentials and set Region from JSON file
+AWS.config.loadFromPath('./aws.json');
+
+// Create S3 service object
+s3 = new AWS.S3({apiVersion: '2006-03-01'});
+
+// Create params JSON for S3.createBucket
+var bucketParams = {
+  Bucket : BUCKETNAME,
+  ACL : 'public-read'
+};
+
+// Create params JSON for S3.setBucketWebsite
+var staticHostParams = {
+  Bucket: BUCKETNAME,
+  WebsiteConfiguration: {
+  ErrorDocument: {
+    Key: 'index.html'
+  },
+  IndexDocument: {
+    Suffix: 'index.html'
+  },
+  }
+};
+
+// Call S3 to create the bucket
+s3.createBucket(bucketParams, function(err, data) {
+  if (err) {
+    console.log("Error", err);
+  } else {
+    console.log("Bucket URL is ", data.Location);
+    // Set the new policy on the newly created bucket
+    s3.putBucketWebsite(staticHostParams, function(err, data) {
+      if (err) {
+        // Display error message
+        console.log("Error", err);
+      } else {
+        // Update the displayed policy for the selected bucket
+        console.log("Success... UPLOADING!", data);
+
+        ///
+        /// After the bucket is created, we upload to it:
+        ///
+        s3FolderUpload(directoryName, credentials, options  , invalidation );
+
+      }
+    });
+  }
+});
