@@ -1,9 +1,12 @@
 import { Card } from "antd";
 import React, { useMemo, useState } from "react";
-import { useContractExistsAtAddress, useContractLoader } from "../../hooks";
+import { useContractExistsAtAddress, useContractLoader } from "eth-hooks";
+
 import Account from "../Account";
 import DisplayVariable from "./DisplayVariable";
 import FunctionForm from "./FunctionForm";
+import Address from "../Address";
+import Balance from "../Balance";
 
 const noContractDisplay = (
   <div>
@@ -53,8 +56,10 @@ export default function Contract({
   show,
   price,
   blockExplorer,
+  chainId,
+  contractConfig,
 }) {
-  const contracts = useContractLoader(provider);
+  const contracts = useContractLoader(provider, contractConfig, chainId);
   let contract;
   if (!customContract) {
     contract = contracts ? contracts[name] : "";
@@ -65,63 +70,61 @@ export default function Contract({
   const address = contract ? contract.address : "";
   const contractIsDeployed = useContractExistsAtAddress(provider, address);
 
-  const displayedContractFunctions = useMemo(
-    () =>
-      contract
-        ? Object.values(contract.interface.functions).filter(
-            fn => fn.type === "function" && !(show && show.indexOf(fn.name) < 0),
-          )
-        : [],
-    [contract, show],
-  );
+  const displayedContractFunctions = useMemo(() => {
+    const results = contract
+      ? Object.entries(contract.interface.functions).filter(
+          fn => fn[1]["type"] === "function" && !(show && show.indexOf(fn[1]["name"]) < 0),
+        )
+      : [];
+    return results;
+  }, [contract, show]);
 
   const [refreshRequired, triggerRefresh] = useState(false);
-  const contractDisplay = displayedContractFunctions.map(fn => {
-    if (isQueryable(fn)) {
-      // If there are no inputs, just display return value
+  const contractDisplay = displayedContractFunctions.map(contractFuncInfo => {
+    const contractFunc =
+      contractFuncInfo[1].stateMutability === "view" || contractFuncInfo[1].stateMutability === "pure"
+        ? contract[contractFuncInfo[0]]
+        : contract.connect(signer)[contractFuncInfo[0]];
+
+    if (typeof contractFunc === "function") {
+      if (isQueryable(contractFuncInfo[1])) {
+        // If there are no inputs, just display return value
+        return (
+          <DisplayVariable
+            key={contractFuncInfo[1].name}
+            contractFunction={contractFunc}
+            functionInfo={contractFuncInfo[1]}
+            refreshRequired={refreshRequired}
+            triggerRefresh={triggerRefresh}
+            blockExplorer={blockExplorer}
+          />
+        );
+      }
+
+      // If there are inputs, display a form to allow users to provide these
       return (
-        <DisplayVariable
-          key={fn.name}
-          contractFunction={contract[fn.name]}
-          functionInfo={fn}
-          refreshRequired={refreshRequired}
+        <FunctionForm
+          key={"FF" + contractFuncInfo[0]}
+          contractFunction={contractFunc}
+          functionInfo={contractFuncInfo[1]}
+          provider={provider}
+          gasPrice={gasPrice}
           triggerRefresh={triggerRefresh}
         />
       );
     }
-    // If there are inputs, display a form to allow users to provide these
-    return (
-      <FunctionForm
-        key={"FF" + fn.name}
-        contractFunction={
-          fn.stateMutability === "view" || fn.stateMutability === "pure"
-            ? contract[fn.name]
-            : contract.connect(signer)[fn.name]
-        }
-        functionInfo={fn}
-        provider={provider}
-        gasPrice={gasPrice}
-        triggerRefresh={triggerRefresh}
-      />
-    );
+    return null;
   });
 
   return (
     <div style={{ margin: "auto", width: "70vw" }}>
       <Card
         title={
-          <div>
+          <div style={{fontSize:24}}>
             {name}
             <div style={{ float: "right" }}>
-              <Account
-                address={address}
-                localProvider={provider}
-                injectedProvider={provider}
-                mainnetProvider={provider}
-                price={price}
-                blockExplorer={blockExplorer}
-              />
-              {account}
+              <Address value={address}/>
+              <Balance address={address} provider={provider} price={price} />
             </div>
           </div>
         }
