@@ -88,15 +88,9 @@ function NetworkDisplay({
         </div>
       );
     }
-  } else {
-    networkDisplay = USE_NETWORK_SELECTOR ? null : (
-      <div style={{ zIndex: -1, position: "absolute", right: 154, top: 28, padding: 16, color: targetNetwork.color }}>
-        {targetNetwork.name}
-      </div>
-    );
   }
 
-  //console.log({ networkDisplay });
+  console.log({ networkDisplay });
 
   return networkDisplay;
 }
