@@ -1,7 +1,8 @@
 import { Badge, Button } from "antd";
-import { useBlockNumber, usePoller } from "eth-hooks";
 import React, { useState } from "react";
+import { useBlockNumber, usePoller } from "eth-hooks";
 // import { WalletOutlined } from '@ant-design/icons';
+
 import Address from "./Address";
 
 export default function Provider(props) {
