@@ -1,7 +1,7 @@
-import { DownloadOutlined, EditOutlined, SendOutlined } from "@ant-design/icons";
 import { Timeline, Typography } from "antd";
 import React from "react";
 import Blockies from "react-blockies";
+import { DownloadOutlined, EditOutlined, SendOutlined } from "@ant-design/icons";
 
 const { Text } = Typography;
 
