@@ -1,6 +1,7 @@
-import { formatEther } from "@ethersproject/units";
 import { useTokenBalance } from "eth-hooks/erc/erc-20/useTokenBalance";
 import React, { useState } from "react";
+
+import { utils } from "ethers";
 
 export default function TokenBalance(props) {
   const [dollarMode, setDollarMode] = useState(true);
@@ -17,7 +18,7 @@ export default function TokenBalance(props) {
   }
 
   if (usingBalance) {
-    const etherBalance = formatEther(usingBalance);
+    const etherBalance = utils.formatEther(usingBalance);
     parseFloat(etherBalance).toFixed(2);
     floatBalance = parseFloat(etherBalance);
   }
