@@ -1,356 +1,58 @@
 // MY INFURA_ID, SWAP IN YOURS FROM https://infura.io/dashboard/ethereum
-export const INFURA_ID = "81fb5d5c883b4aa7bdccd77f0794c40d";
+export const INFURA_ID = "460f40a260564ac4a4f4b3fffb032dad";
 
 // MY ETHERSCAN_ID, SWAP IN YOURS FROM https://etherscan.io/myapikey
-export const ETHERSCAN_KEY = "PSW8C433Q667DVEX5BCRMGNAH9FSGFZ7Q8";
+export const ETHERSCAN_KEY = "DNXJA8RX2Q3VZ4URQIWP7Z68CJXQZSC6AW";
 
 // BLOCKNATIVE ID FOR Notify.js:
 export const BLOCKNATIVE_DAPPID = "0b58206a-f3c0-4701-a62f-73c7243e8c77";
 
-// EXTERNAL CONTRACTS
-
-export const DAI_ADDRESS = "0x6B175474E89094C44Da98b954EedeAC495271d0F";
-
-export const DAI_ABI = [
-  {
-    inputs: [{ internalType: "uint256", name: "chainId_", type: "uint256" }],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "constructor",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: "address", name: "src", type: "address" },
-      { indexed: true, internalType: "address", name: "guy", type: "address" },
-      { indexed: false, internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "Approval",
-    type: "event",
-  },
-  {
-    anonymous: true,
-    inputs: [
-      { indexed: true, internalType: "bytes4", name: "sig", type: "bytes4" },
-      { indexed: true, internalType: "address", name: "usr", type: "address" },
-      { indexed: true, internalType: "bytes32", name: "arg1", type: "bytes32" },
-      { indexed: true, internalType: "bytes32", name: "arg2", type: "bytes32" },
-      { indexed: false, internalType: "bytes", name: "data", type: "bytes" },
-    ],
-    name: "LogNote",
-    type: "event",
-  },
-  {
-    anonymous: false,
-    inputs: [
-      { indexed: true, internalType: "address", name: "src", type: "address" },
-      { indexed: true, internalType: "address", name: "dst", type: "address" },
-      { indexed: false, internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "Transfer",
-    type: "event",
-  },
-  {
-    constant: true,
-    inputs: [],
-    name: "DOMAIN_SEPARATOR",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [],
-    name: "PERMIT_TYPEHASH",
-    outputs: [{ internalType: "bytes32", name: "", type: "bytes32" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [
-      { internalType: "address", name: "", type: "address" },
-      { internalType: "address", name: "", type: "address" },
-    ],
-    name: "allowance",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [
-      { internalType: "address", name: "usr", type: "address" },
-      { internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "approve",
-    outputs: [{ internalType: "bool", name: "", type: "bool" }],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [{ internalType: "address", name: "", type: "address" }],
-    name: "balanceOf",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [
-      { internalType: "address", name: "usr", type: "address" },
-      { internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "burn",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [],
-    name: "decimals",
-    outputs: [{ internalType: "uint8", name: "", type: "uint8" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [{ internalType: "address", name: "guy", type: "address" }],
-    name: "deny",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [
-      { internalType: "address", name: "usr", type: "address" },
-      { internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "mint",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [
-      { internalType: "address", name: "src", type: "address" },
-      { internalType: "address", name: "dst", type: "address" },
-      { internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "move",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [],
-    name: "name",
-    outputs: [{ internalType: "string", name: "", type: "string" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [{ internalType: "address", name: "", type: "address" }],
-    name: "nonces",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [
-      { internalType: "address", name: "holder", type: "address" },
-      { internalType: "address", name: "spender", type: "address" },
-      { internalType: "uint256", name: "nonce", type: "uint256" },
-      { internalType: "uint256", name: "expiry", type: "uint256" },
-      { internalType: "bool", name: "allowed", type: "bool" },
-      { internalType: "uint8", name: "v", type: "uint8" },
-      { internalType: "bytes32", name: "r", type: "bytes32" },
-      { internalType: "bytes32", name: "s", type: "bytes32" },
-    ],
-    name: "permit",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [
-      { internalType: "address", name: "usr", type: "address" },
-      { internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "pull",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [
-      { internalType: "address", name: "usr", type: "address" },
-      { internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "push",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [{ internalType: "address", name: "guy", type: "address" }],
-    name: "rely",
-    outputs: [],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [],
-    name: "symbol",
-    outputs: [{ internalType: "string", name: "", type: "string" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [],
-    name: "totalSupply",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [
-      { internalType: "address", name: "dst", type: "address" },
-      { internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "transfer",
-    outputs: [{ internalType: "bool", name: "", type: "bool" }],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: false,
-    inputs: [
-      { internalType: "address", name: "src", type: "address" },
-      { internalType: "address", name: "dst", type: "address" },
-      { internalType: "uint256", name: "wad", type: "uint256" },
-    ],
-    name: "transferFrom",
-    outputs: [{ internalType: "bool", name: "", type: "bool" }],
-    payable: false,
-    stateMutability: "nonpayable",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [],
-    name: "version",
-    outputs: [{ internalType: "string", name: "", type: "string" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-  {
-    constant: true,
-    inputs: [{ internalType: "address", name: "", type: "address" }],
-    name: "wards",
-    outputs: [{ internalType: "uint256", name: "", type: "uint256" }],
-    payable: false,
-    stateMutability: "view",
-    type: "function",
-  },
-];
-
-export const NETWORK = chainId => {
-  for (const n in NETWORKS) {
-    if (NETWORKS[n].chainId == chainId) {
-      return NETWORKS[n];
-    }
-  }
-};
+export const ALCHEMY_KEY = "oKxs-03sij-U_N0iOlrSsZFr29-IqbuF";
 
 export const NETWORKS = {
-  ethereum: {
-    name: "ethereum",
-    color: "#ceb0fa",
+  localhost: {
+    name: "localhost",
+    color: "#666666",
+    chainId: 31337,
+    blockExplorer: "",
+    rpcUrl: "http://" + (global.window ? window.location.hostname : "localhost") + ":8545",
+  },
+  mainnet: {
+    name: "mainnet",
+    color: "#ff8b9e",
     chainId: 1,
-    price: "uniswap",
     rpcUrl: `https://mainnet.infura.io/v3/${INFURA_ID}`,
     blockExplorer: "https://etherscan.io/",
-  },
-  optimism: {
-    name: "optimism",
-    color: "#f01a37",
-    price: "uniswap",
-    chainId: 10,
-    blockExplorer: "https://optimistic.etherscan.io/",
-    rpcUrl: `https://mainnet.optimism.io`,
-    //gasPrice: 1000000,
-  },
-  arbitrum: {
-    name: "arbitrum",
-    color: "#50a0ea",
-    price: "uniswap",
-    chainId: 42161,
-    blockExplorer: "https://arbiscan.io/",
-    rpcUrl: `https://arb1.arbitrum.io/rpc`,
-    //gasPrice: 1000000000,// TODO ASK RPC
   },
   gnosis: {
     name: "gnosis",
     color: "#48a9a6",
     chainId: 100,
     price: 1,
-    nativeCurrency: {
-      name: "xDAI",
-      symbol: "xDAI",
-      decimals: 18,
-    },
-    gasPrice: 1000000000,
-    rpcUrl: "https://rpc.gnosischain.com/",
+    gasPrice: 4000000000,
+    rpcUrl: "https://dai.poa.network",
     faucet: "https://xdai-faucet.top/",
     blockExplorer: "https://blockscout.com/poa/xdai/",
   },
   polygon: {
     name: "polygon",
     color: "#2bbdf7",
-    price: "uniswap:0x7D1AfA7B718fb893dB30A3aBc0Cfc608AaCfeBB0",
     chainId: 137,
-    rpcUrl: "https://polygon-rpc.com",
-    faucet: "https://faucet.matic.network/",
-    blockExplorer: "https://explorer-mainnet.maticvigil.com//",
+    price: 1,
+    gasPrice: 61000000000,
+    rpcUrl: "https://polygon-rpc.com/",
+    blockExplorer: "https://polygonscan.com/",
+  },
+  optimism: {
+    name: "optimism",
+    color: "#f01a37",
+    chainId: 10,
+    blockExplorer: "https://optimistic.etherscan.io/",
+    rpcUrl: `https://mainnet.optimism.io`,
   },
   kovan: {
     name: "kovan",
     color: "#7003DD",
-    price: "uniswap",
     chainId: 42,
     rpcUrl: `https://kovan.infura.io/v3/${INFURA_ID}`,
     blockExplorer: "https://kovan.etherscan.io/",
@@ -359,7 +61,6 @@ export const NETWORKS = {
   rinkeby: {
     name: "rinkeby",
     color: "#e0d068",
-    price: "uniswap",
     chainId: 4,
     rpcUrl: `https://rinkeby.infura.io/v3/${INFURA_ID}`,
     faucet: "https://faucet.rinkeby.io/",
@@ -367,8 +68,7 @@ export const NETWORKS = {
   },
   ropsten: {
     name: "ropsten",
-    color: "#ff4a8d",
-    price: "uniswap",
+    color: "#F60D09",
     chainId: 3,
     faucet: "https://faucet.ropsten.be/",
     blockExplorer: "https://ropsten.etherscan.io/",
@@ -377,28 +77,96 @@ export const NETWORKS = {
   goerli: {
     name: "goerli",
     color: "#0975F6",
-    price: "uniswap",
     chainId: 5,
     faucet: "https://goerli-faucet.slock.it/",
     blockExplorer: "https://goerli.etherscan.io/",
     rpcUrl: `https://goerli.infura.io/v3/${INFURA_ID}`,
   },
-  localhost: {
-    name: "localhost",
-    color: "#666666",
-    price: "uniswap", // use mainnet eth price for localhost
-    chainId: 31337,
-    blockExplorer: "",
-    rpcUrl: "http://localhost:8545",
-  },
   mumbai: {
     name: "mumbai",
     color: "#92D9FA",
     chainId: 80001,
+    price: 1,
     gasPrice: 1000000000,
     rpcUrl: "https://rpc-mumbai.maticvigil.com",
-    faucet: "https://faucet.matic.network/",
-    blockExplorer: "https://mumbai-explorer.matic.today/",
+    faucet: "https://faucet.polygon.technology/",
+    blockExplorer: "https://mumbai.polygonscan.com/",
+  },
+  localArbitrum: {
+    name: "localArbitrum",
+    color: "#50a0ea",
+    chainId: 153869338190755,
+    blockExplorer: "",
+    rpcUrl: `http://localhost:8547`,
+  },
+  localArbitrumL1: {
+    name: "localArbitrumL1",
+    color: "#50a0ea",
+    chainId: 44010,
+    blockExplorer: "",
+    rpcUrl: `http://localhost:7545`,
+  },
+  rinkebyArbitrum: {
+    name: "Arbitrum Testnet",
+    color: "#50a0ea",
+    chainId: 421611,
+    blockExplorer: "https://rinkeby-explorer.arbitrum.io/#/",
+    rpcUrl: `https://rinkeby.arbitrum.io/rpc`,
+  },
+  arbitrum: {
+    name: "Arbitrum",
+    color: "#50a0ea",
+    chainId: 42161,
+    blockExplorer: "https://explorer.arbitrum.io/#/",
+    rpcUrl: `https://arb1.arbitrum.io/rpc`,
+    gasPrice: 0,
+  },
+  localOptimismL1: {
+    name: "localOptimismL1",
+    color: "#f01a37",
+    chainId: 31337,
+    blockExplorer: "",
+    rpcUrl: "http://" + (global.window ? window.location.hostname : "localhost") + ":9545",
+  },
+  localOptimism: {
+    name: "localOptimism",
+    color: "#f01a37",
+    chainId: 420,
+    blockExplorer: "",
+    rpcUrl: "http://" + (global.window ? window.location.hostname : "localhost") + ":8545",
+    gasPrice: 0,
+  },
+  kovanOptimism: {
+    name: "kovanOptimism",
+    color: "#f01a37",
+    chainId: 69,
+    blockExplorer: "https://kovan-optimistic.etherscan.io/",
+    rpcUrl: `https://kovan.optimism.io`,
+    gasPrice: 0,
+  },
+  localAvalanche: {
+    name: "localAvalanche",
+    color: "#666666",
+    chainId: 43112,
+    blockExplorer: "",
+    rpcUrl: `http://localhost:9650/ext/bc/C/rpc`,
+    gasPrice: 225000000000,
+  },
+  fujiAvalanche: {
+    name: "fujiAvalanche",
+    color: "#666666",
+    chainId: 43113,
+    blockExplorer: "https://cchain.explorer.avax-test.network/",
+    rpcUrl: `https://api.avax-test.network/ext/bc/C/rpc`,
+    gasPrice: 225000000000,
+  },
+  mainnetAvalanche: {
+    name: "mainnetAvalanche",
+    color: "#666666",
+    chainId: 43114,
+    blockExplorer: "https://cchain.explorer.avax.network/",
+    rpcUrl: `https://api.avax.network/ext/bc/C/rpc`,
+    gasPrice: 225000000000,
   },
   testnetHarmony: {
     name: "testnetHarmony",
@@ -407,15 +175,38 @@ export const NETWORKS = {
     blockExplorer: "https://explorer.pops.one/",
     rpcUrl: `https://api.s0.b.hmny.io`,
     gasPrice: 1000000000,
-    token: "ONE",
   },
-  harmony: {
-    name: "harmony",
+  mainnetHarmony: {
+    name: "mainnetHarmony",
     color: "#00b0ef",
     chainId: 1666600000,
     blockExplorer: "https://explorer.harmony.one/",
     rpcUrl: `https://api.harmony.one`,
     gasPrice: 1000000000,
-    token: "ONE",
   },
+  fantom: {
+    name: "fantom",
+    color: "#1969ff",
+    chainId: 250,
+    blockExplorer: "https://ftmscan.com/",
+    rpcUrl: `https://rpcapi.fantom.network`,
+    gasPrice: 1000000000,
+  },
+  testnetFantom: {
+    name: "testnetFantom",
+    color: "#1969ff",
+    chainId: 4002,
+    blockExplorer: "https://testnet.ftmscan.com/",
+    rpcUrl: `https://rpc.testnet.fantom.network`,
+    gasPrice: 1000000000,
+    faucet: "https://faucet.fantom.network/",
+  },
+};
+
+export const NETWORK = chainId => {
+  for (const n in NETWORKS) {
+    if (NETWORKS[n].chainId === chainId) {
+      return NETWORKS[n];
+    }
+  }
 };
