@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 
-/*
+/**
   ~ What it does? ~
 
   Gets a tokenlist (see more at https://tokenlists.org/), returning the .tokens only
@@ -13,7 +13,7 @@ import { useEffect, useState } from "react";
   ~ Features ~
 
   - Optional - specify chainId to filter by chainId
-*/
+**/
 
 const useTokenList = (tokenListUri, chainId) => {
   const [tokenList, setTokenList] = useState([]);
@@ -22,23 +22,23 @@ const useTokenList = (tokenListUri, chainId) => {
 
   useEffect(() => {
     const getTokenList = async () => {
-      try {
-        const tokenList = await fetch(_tokenListUri);
-        const tokenListJson = await tokenList.json();
-        let _tokenList;
+        try {
+          const tokenList = await fetch(_tokenListUri);
+          const tokenListJson = await tokenList.json();
+          let _tokenList;
 
-        if (chainId) {
-          _tokenList = tokenListJson.tokens.filter(function (t) {
-            return t.chainId === chainId;
-          });
-        } else {
-          _tokenList = tokenListJson;
+          if (chainId) {
+            _tokenList = tokenListJson.tokens.filter(function (t) {
+              return t.chainId === chainId;
+            });
+          } else {
+            _tokenList = tokenListJson;
+          }
+
+          setTokenList(_tokenList.tokens);
+        } catch (e) {
+          console.log(e);
         }
-
-        setTokenList(_tokenList.tokens);
-      } catch (e) {
-        console.log(e);
-      }
     };
     getTokenList();
   }, [tokenListUri]);
