@@ -1,17 +1,14 @@
-/* eslint-disable jsx-a11y/accessible-emoji */
-
-import { SyncOutlined } from "@ant-design/icons";
-import { formatEther, parseEther } from "@ethersproject/units";
-import { Button, Card, DatePicker, Divider, Input, List, Progress, Slider, Spin, Switch } from "antd";
+import { Button, Card, DatePicker, Divider, Input, Progress, Slider, Spin, Switch } from "antd";
 import React, { useState } from "react";
-import { Address, Balance } from "../components";
+import { utils } from "ethers";
+import { SyncOutlined } from "@ant-design/icons";
+
+import { Address, Balance, Events } from "../components";
 
 export default function ExampleUI({
   purpose,
-  setPurposeEvents,
   address,
   mainnetProvider,
-  userProvider,
   localProvider,
   yourLocalBalance,
   price,
@@ -37,13 +34,30 @@ export default function ExampleUI({
             }}
           />
           <Button
-            onClick={() => {
-              console.log("newPurpose", newPurpose);
+            style={{ marginTop: 8 }}
+            onClick={async () => {
               /* look how you call setPurpose on your contract: */
-              tx(writeContracts.YourContract.setPurpose(newPurpose));
+              /* notice how you pass a call back for tx updates too */
+              const result = tx(writeContracts.YourContract.setPurpose(newPurpose), update => {
+                console.log("📡 Transaction Update:", update);
+                if (update && (update.status === "confirmed" || update.status === 1)) {
+                  console.log(" 🍾 Transaction " + update.hash + " finished!");
+                  console.log(
+                    " ⛽️ " +
+                      update.gasUsed +
+                      "/" +
+                      (update.gasLimit || update.gas) +
+                      " @ " +
+                      parseFloat(update.gasPrice) / 1000000000 +
+                      " gwei",
+                  );
+                }
+              });
+              console.log("awaiting metamask/web3 confirm result...", result);
+              console.log(await result);
             }}
           >
-            Set Purpose
+            Set Purpose!
           </Button>
         </div>
         <Divider />
@@ -57,20 +71,20 @@ export default function ExampleUI({
           fontSize={16}
         />
         <Divider />
-        {/* use formatEther to display a BigNumber: */}
-        <h2>Your Balance: {yourLocalBalance ? formatEther(yourLocalBalance) : "..."}</h2>
+        {/* use utils.formatEther to display a BigNumber: */}
+        <h2>Your Balance: {yourLocalBalance ? utils.formatEther(yourLocalBalance) : "..."}</h2>
         <div>OR</div>
         <Balance address={address} provider={localProvider} price={price} />
         <Divider />
         <div>🐳 Example Whale Balance:</div>
-        <Balance balance={parseEther("1000")} provider={localProvider} price={price} />
+        <Balance balance={utils.parseEther("1000")} provider={localProvider} price={price} />
         <Divider />
-        {/* use formatEther to display a BigNumber: */}
-        <h2>Your Balance: {yourLocalBalance ? formatEther(yourLocalBalance) : "..."}</h2>
+        {/* use utils.formatEther to display a BigNumber: */}
+        <h2>Your Balance: {yourLocalBalance ? utils.formatEther(yourLocalBalance) : "..."}</h2>
         <Divider />
         Your Contract Address:
         <Address
-          address={readContracts ? readContracts.YourContract.address : readContracts}
+          address={readContracts && readContracts.YourContract ? readContracts.YourContract.address : null}
           ensProvider={mainnetProvider}
           fontSize={16}
         />
@@ -94,7 +108,7 @@ export default function ExampleUI({
             */
               tx({
                 to: writeContracts.YourContract.address,
-                value: parseEther("0.001"),
+                value: utils.parseEther("0.001"),
               });
               /* this should throw an error about "no fallback nor receive function" until you add it */
             }}
@@ -108,7 +122,7 @@ export default function ExampleUI({
               /* look how we call setPurpose AND send some value along */
               tx(
                 writeContracts.YourContract.setPurpose("💵 Paying for this one!", {
-                  value: parseEther("0.001"),
+                  value: utils.parseEther("0.001"),
                 }),
               );
               /* this will fail until you make the setPurpose function payable */
@@ -123,7 +137,7 @@ export default function ExampleUI({
               /* you can also just craft a transaction and send it to the tx() transactor */
               tx({
                 to: writeContracts.YourContract.address,
-                value: parseEther("0.001"),
+                value: utils.parseEther("0.001"),
                 data: writeContracts.YourContract.interface.encodeFunctionData("setPurpose(string)", [
                   "🤓 Whoa so 1337!",
                 ]),
@@ -140,21 +154,14 @@ export default function ExampleUI({
         📑 Maybe display a list of events?
           (uncomment the event and emit line in YourContract.sol! )
       */}
-      <div style={{ width: 600, margin: "auto", marginTop: 32, paddingBottom: 32 }}>
-        <h2>Events:</h2>
-        <List
-          bordered
-          dataSource={setPurposeEvents}
-          renderItem={item => {
-            return (
-              <List.Item key={item.blockNumber + "_" + item.sender + "_" + item.purpose}>
-                <Address address={item[0]} ensProvider={mainnetProvider} fontSize={16} /> =&gt
-                {item[1]}
-              </List.Item>
-            );
-          }}
-        />
-      </div>
+      <Events
+        contracts={readContracts}
+        contractName="YourContract"
+        eventName="SetPurpose"
+        localProvider={localProvider}
+        mainnetProvider={mainnetProvider}
+        startBlock={1}
+      />
 
       <div style={{ width: 600, margin: "auto", marginTop: 32, paddingBottom: 256 }}>
         <Card>
