@@ -1,11 +1,13 @@
-/* eslint-disable jsx-a11y/accessible-emoji */
-import { gql, useQuery } from "@apollo/client";
-import { Button, Input, Table, Typography } from "antd";
 import "antd/dist/antd.css";
-import GraphiQL from "graphiql";
 import "graphiql/graphiql.min.css";
-import fetch from "isomorphic-fetch";
+
 import React, { useState } from "react";
+import { Button, Input, Table, Typography } from "antd";
+import { gql, useQuery } from "@apollo/client";
+import GraphiQL from "graphiql";
+import fetch from "isomorphic-fetch";
+
+
 import { Address } from "../components";
 
 const highlight = {
@@ -81,7 +83,7 @@ function Subgraph(props) {
       </div>
       <div style={{ margin: "auto", marginTop: 32 }}>
         Instead, you can use{" "}
-        <a href="https://thegraph.com/docs/introduction" target="_blank" rel="noopener noreferrer">
+        <a href="https://thegraph.com/docs/about/introduction" target="_blank" rel="noopener noreferrer">
           The Graph
         </a>{" "}
         with 🏗 scaffold-eth (
@@ -95,7 +97,7 @@ function Subgraph(props) {
         <span style={{ marginRight: 8 }}>🚮</span>
         Clean up previous data:
         <span className="highlight" style={highlight}>
-          rm -rf docker/graph-node/data/
+          yarn clean-graph-node
         </span>
       </div>
 
@@ -103,7 +105,7 @@ function Subgraph(props) {
         <span style={{ marginRight: 8 }}>📡</span>
         Spin up a local graph node by running
         <span className="highlight" style={highlight}>
-          yarn graph-run-node
+          yarn run-graph-node
         </span>
         <span style={{ marginLeft: 4 }}>
           {" "}
